+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Add support for purchase prices
 * Add support for game editions (https://git.kucharczyk.xyz/lukas/timetracker/issues/28)
 
 ## 1.0.1 / 2023-01-30 22:17+01:00